src/backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-otp
qrcode
python-decouple
psycopg2-binary
